requirements.txt
colorama==0.4.6
keyboard==0.13.5
MouseInfo==0.1.3
numpy==1.26.2
opencv-python==4.8.1.78
Pillow==10.1.0
pydirectinput-rgx==2.1.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pytweening==1.0.7
